LayoutTests/js/arrowfunction-tdz-expected.txt

Tests for ES6 arrow function test tdz

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isReferenceError is true
PASS successfullyParsed is true

TEST COMPLETE
